Provide standardized validation for location types
This validation is for when the API requests mandates a particular
type of location such as `addressLocation` or `facilityLocation`.  In
this case, it can be used as the following:

```java
   @ValidLocationSubtype(anyOf = {ValidLocationSubtype.LocationSubtype.ADDRESS, ValidLocationSubtype.LocationSubtype.FACILITY})
   private final LocationTO addressOrFacilityLocationType;

```

Note that the annotation provides a minimum validation and permits
other fields to be present as well.  Accordingly, a geo coordinate
with an address would count as both a `GEO_COORDINATE` and an
`ADDRESS` location at the same time and would pass the restriction
from the example.

In the rare case that `and` restrictions are needed, then the
annotation can be repeated.  It will still not prevent "additional"
fields but can be useful to ensure that you have a facility with a UN
Location Code (to future proof in case we get facility code lists that
do not depend on UN Location codes).  As an example:

```java
     // Either UN Location Code OR facility (or both)
     @ValidLocationSubtype(anyOf = {ValidLocationSubtype.LocationSubtype.UN_LOCATION, ValidLocationSubtype.LocationSubtype.FACILITY})
     private final LocationTO poteLocation;

     // UN Location code AND facility
     @ValidLocationSubtype(anyOf = ValidLocationSubtype.LocationSubtype.UN_LOCATION)
     @ValidLocationSubtype(anyOf = ValidLocationSubtype.LocationSubtype.FACILITY)
     private final LocationTO portWithTerminalLocation;
```

Signed-off-by: Niels Thykier <nt@asseco.dk>